backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0

# Real-time communication
python-socketio>=5.0.0

# Scheduler
apscheduler>=3.10.0

# HTTP client (for exchange rates)
httpx>=0.25.0
aiohttp>=3.8.0

# Validation and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Date/time handling
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
pylint>=3.0.0

